# 8310160: Make GC APIs for handling archive heap objects agnostic of GC policy