--- a/src/hotspot/share/gc/g1/g1CollectedHeap.hpp
+++ b/src/hotspot/share/gc/g1/g1CollectedHeap.hpp
@@ -711,17 +711,19 @@ public:
   // not be same as the preferred address.
   // This API is only used for allocating heap space for the archived heap objects
   // in the CDS archive.
-  HeapWord* alloc_archive_region(size_t word_size, HeapWord* preferred_addr);
+  HeapWord* alloc_archive_space(size_t word_size, HeapWord* preferred_addr) override;
 
-  // Populate the G1BlockOffsetTablePart for archived regions with the given
-  // memory range.
-  void populate_archive_regions_bot_part(MemRegion range);
+  // Populate the G1BlockOffsetTableParts for the archived regions in the given range.
+  // That ensures fast G1BlockOffsetTablePart::block_start operations for any given
+  // address within the archive regions when trying to find start of an object
+  // (e.g. during card table scanning).
+  void fixup_archive_space(MemRegion range) override;
 
   // For the specified range, uncommit the containing G1 regions
-  // which had been allocated by alloc_archive_regions. This should be called
+  // which had been allocated by alloc_archive_space. This should be called
   // at JVM init time if the archive heap's contents cannot be used (e.g., if
   // CRC check fails).
-  void dealloc_archive_regions(MemRegion range);
+  void handle_archive_space_failure(MemRegion range) override;
 
 private:
 
@@ -1044,10 +1046,6 @@ public:
 
   inline G1HeapRegionAttr region_attr(const void* obj) const;
   inline G1HeapRegionAttr region_attr(uint idx) const;
-
-  MemRegion reserved() const {
-    return _hrm.reserved();
-  }
 
   bool is_in_reserved(const void* addr) const {
     return reserved().contains(addr);