--- a/src/hotspot/share/gc/g1/g1CollectedHeap.hpp
+++ b/src/hotspot/share/gc/g1/g1CollectedHeap.hpp
@@ -1048,7 +1048,7 @@ public:
   inline G1HeapRegionAttr region_attr(uint idx) const;
 
   bool is_in_reserved(const void* addr) const {
-    return reserved().contains(addr);
+    return reserved_region().contains(addr);
   }
 
   G1CardTable* card_table() const {