@ashu-mehra 2023-06-21
My whole idea behind this is to have common GC APIs that the CDS can use to manage archive heap. As long as collector internals are properly abstracted out and hidden behind the APIs, CDS wouldn't need to worry about what the individual collectors do internally.
I agree the ideal scenario would be to use object allocation APIs but considering the issues @iklam pointed out, it doesn't look like a viable approach. @fisk do you agree with this assessment? Do you have any other alternative approach in mind?