@iklam 2023-06-23
Based on discussion with @fisk, I created an RFE for investigating his idea. Please see https://bugs.openjdk.org/browse/JDK-8310823

It seems quite promising to me, and will greatly reduce (or eliminate) the interface between CDS and the collectors.

I would like to keep the current performance as possible, so I am leaning towards having an API for CDS to tell the collector about preferred location of the archived objects, to allow mmap'ing and reduce/avoid relocation. But such an hinting API will be much smaller than the ones proposed by this PR.

(We'd also need an reverse API for the collector to tell CDS what the preferred address would be, during CDS dump time).