@iklam 2023-07-05
> Based on discussion with @fisk, I created an RFE for investigating his idea. Please see https://bugs.openjdk.org/browse/JDK-8310823
> 
> It seems quite promising to me, and will greatly reduce (or eliminate) the interface between CDS and the collectors.
> 
> I would like to keep the current performance as possible, so I am leaning towards having an API for CDS to tell the collector about preferred location of the archived objects, to allow mmap'ing and reduce/avoid relocation. But such an hinting API will be much smaller than the ones proposed by this PR.
> 
> (We'd also need an reverse API for the collector to tell CDS what the preferred address would be, during CDS dump time).

I've done a very simple (and rough) implementation of @fisk 's idea, without any optimizations. Every relocation is done via a hashtable lookup. For the default CDS archive, this makes about 48000 relocations on start-up.

https://github.com/openjdk/jdk/compare/master...iklam:jdk:8310823-materialize-cds-heap-with-regular-alloc?expand=1

```
$ perf stat -r 40 java --version > /dev/null
          0.015065 +- 0.000228 seconds time elapsed  ( +-  1.51% )

$ perf stat -r 40 java -XX:+NewArchiveHeapLoading --version > /dev/null
          0.020598 +- 0.000215 seconds time elapsed  ( +-  1.04% )

$ perf stat -r 40 java -XX:+UseSerialGC --version > /dev/null
          0.013929 +- 0.000229 seconds time elapsed  ( +-  1.64% )

$ perf stat -r 40 java -XX:+UseSerialGC -XX:+NewArchiveHeapLoading --version > /dev/null
          0.019107 +- 0.000222 seconds time elapsed  ( +-  1.16% )
```

The cost of the individual object allocation and relocation is about 5ms.

So far the slow down doesn't seem too outrages.

My next step is to optimize the relocation code to see how much faster it can get.

I noticed that the objects in `ArchiveHeapLoader::newcode_runtime_allocate_objects()` are allocated in at least two contiguous blocks, due to TLAB overflow.